requirements.txt
numpy=1.15.4
pytorch=1.0.1
torchvision=0.2.2
pillow=5.3.0